refactor(instruction): extract typed single-instruction discriminator helper
